Database updates should now be contained in files called 'something_xxx.sql' where xxx is the version of the database that the file upgrades to. xxx can be as many numbers as is needed e.g. 0004 or 03 or 8 etc. 'something' can be any name at all. Any files with a .sql extension that don't conform to this naming convension will be ignored.
The update script will now only download .sql files that have a valid name (with numeric version) and if the file is a later version than the current database version.

SQL files will be executed directly from the updates directory and will not be renamed into a file in the main tree. This is done since these are transitory files and shouldn't be stored once they've been applied as the end users may end up with a large number of them.

If an error occurs during processing of a SQL file, the updates will stop there and the next time an update is run it will carry on from the file that failed. This will allow that file to be patched to fix the problem and the database to be kept in a known state rather than blindly applying updates when to a damaged database.

NOTE: It is important with the 1.05 release to ensure that the SQL update file db_update_0002.sql has the correct version inserted as this will be used for future updates. This number should be the number of the last file in the 1.05 release.

NOTE: This code is UNTESTED, it really needs testing before going live. Parts have been tested in isolation but a full test has not been done.