Investigated problems with downloading info from Lovefilm (Refs #76).
Removed output of all SQL statements to the log until multiple debug levels are defined as this is swamping the information present.

Added the output of Directors/Actors/Genres to the logfile for movie information. This helps the user to identify if it's working properly.

Added a function to change the word order of titles beginning with 'The' to accomodate the fact that DVDplanet title files like this: 'Abyss, The'.