Added new code (and moved a couple of functions around within the various libraries) to provide a link to Wikipedia as per Dejunai's suggestions.
However, I have put the music link to first link to the ARTIST if it is unique, and then the ALBUM if not. I think this has a better chance of directing the user to a useful page as the artist pages generally list the albums released by that artist... whereas album names can be very misleading (eg: "18", "Play", "Deeper", etc).

Refs #87.