Added two new functions to determine specific details about the PHP environment.
php_cli_location() - returns the location (path and filename) of the PHP executable that is used for CLI execution).

php_ini_location() - returns the location of the "php.ini" file that is currently in use.

Also modified the run_background() function to use these new functions for any jobs that it starts in the background.

Refs #79.