Added a sanity check to the procedures responsible for manipulating the user's history while browsing media.
The check makes sure that the stack is never empty (which only happens when browsing via the PC and you click the "browser" back, rather than the navigation button).

This should also prevent manipulation of the URL to remove the constraint to view media that is above of your user's maximum rating.