Fixed the bug that meant old movie details were not being cleared when the user manually clicks the "Update details" button on the config video details page.
Removed the lovefilm movie information parser as it needed updating due to a redesign of the lovefilm site, and the effort to recode it just wasn't worth it due to the poor quality of information/searching available at lovefilm.

Eg: A search for "Star Wars" does not return the original Start Wars (IV. A New Hope) - t returns matches such as "lazing Saddles".. wtf?).