The top-level pages for choosing a music/movie/photo category used !empty() to determine if a category had been selected, and "All Categories" used a category ID of -1 to denote a special case.
However, sub-categories now use negative numbers (why?) and the "All categories" ID was changed to zero.

The empty() function considers zero to be empty and therefore zero could not be used to denote a category ID. I have changed the use of !empty() to isset() instead. This achieves the desired behaviour.

fixes #98.