Added option to specify an iTunes Music Library. If an iTunes library exists it is now parsed and any Playlists will be added to SwissCenter.
There is still an issue with character encoding. The iTunes library is in UTF-8 which causes problems when querying MySQL which is latin1. Any tracks/artists containing special characters are not added to the playlist (ie. Björk). This also applies to the playlist name (ie. "90's Music" is saved as "90â€™s Music").

refs #30