Added a function to return the correct text for the Filter link (so that it shows when a filter is active).
Also ensured that the filter is cleared when the user visits the main menu (it's slightly unintuitive that setting a filter in the movies section will also enforce the filter in the music/photos sections).

Added more functions for restricting media by the number of viewings, or viewing state (eg: played, not played, etc).