Included windows executables ''bgrun'' (to run a program in the background, returning control to the PHP script immediately) and ''wget'' (to run SwissCenter scripts via the webserver and remove the reliance on having a CLI version of PHP installedand configured correctly).
The run_background() function now uses ''bgrun'' to execute a command in the background under windows, and this appears to work perfectly. 

Scheduling media refreshes under windows (non-simiese) now uses ''at'' correctly (quotes are in the correct places!).

The run_foreground() function is no longer required and has been removed. 

As a result of the move from a CLI PHP to using wget, the code checking the Swisscenter installation (on the main config screen) is much cleaner and easier to follow.

fixes #43