The GMT offset was not being stored within the session because session files can get locked when multiple pages are accessing it (when music is playing there are two concurrent pages being requested - the "Now Playing" screen and the streaming page). It is also worth noting that the session variables only gets written when the page ends (which is a long time for the streaming page)
To get around this, the GMT offset is saved to the database instead of the session. 

I've also added a "modified" column to both the sys and user preferences tables which gets updated with the time that the preference was last updated. This is then used for the GMT offset to ensure that we only use it if it was calculated less than 24 horus ago so that changes in timezone (eg: in/out of DST) are picked up.

Refs #128.