Added new media type 'DVD Video'. This can only be used if the media player supports playback of DVD images (ie. Popcorn Hour and other NMT).
Supported formats are ISO, IMG, and folders containing a valid VIDEO_TS folder.

For a NMT to be able to play a DVD via HTTP it is necessary to setup SMB/NFS network shares for the folders containing DVD images. This share must then be referenced with the Media Location defined in SwissCenter.