Added new parser for themoviedb.org which also downloads fanart thumbnails to be used as 'Themes'. This parser also recognises IMDB id's as part of the filename/title of media. ie. 'Halloween [tt0077651]'.
Updated the TheTVDB parser to download fanart thumbnails to be used as 'Themes'.

Added a 'Themes' section to the Config to select from available fanart to be used as page background images. Thumbnails for all available fanart for a movie/TV series are downloaded by the parsers to the cache folder. Full size images are only downloaded and processed (flipped, greyscale) for selected fanart.

Refs #176