Added ability to browse/search Apple movie trailers.
The popular RSS and XML feeds that other Apple movie trailer apps use were found to contain only a limited number of recent trailers. We use the JSON requests that are used at http://www.apple.com/trailers/. If the trailer can be browsed/searched for at the site then it is also available in SwissCenter.

Fixes #197